store/requirements.txt
flask
pymongo
redis
pillow